pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.3.7",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "wikipedia>=1.4.0",
]
